Move TimeZoneUtil to the Quartz namespace
FindTimeZoneById is how a trigger built with InTimeZone(...) comes back
out of a job store, and the wall-clock GetUtcOffset overload is the
scheduler-wide daylight saving policy - scheduling API, not a utility.
The type moves next to the builders whose behavior it defines, and the
test moves out of Utils/ with it.

No SimpleTypeLoadHelper shim entry: TimeZoneUtil is a static helper that
is called, never a type a configuration string names for instantiation.
