+++ docs/documentation/quartz-4.x/tutorial/recurrencetrigger.md
@@ -114,7 +114,7 @@ using the builder's `InTimeZone` method:
 ITrigger trigger = TriggerBuilder.Create()
     .WithIdentity("trigger1", "group1")
     .WithRecurrenceSchedule("FREQ=MONTHLY;BYDAY=2MO", b => b
-        .InTimeZone(Quartz.Util.TimeZoneUtil.FindTimeZoneById("Eastern Standard Time")))
+        .InTimeZone(TimeZoneUtil.FindTimeZoneById("Eastern Standard Time")))
     .StartNow()
     .Build();
 ```
