+++ src/Quartz.Tests.Unit/Impl/Calendar/MonthlyCalendarTest.cs
@@ -21,7 +21,6 @@
 
 using Quartz.Impl.Calendar;
 using Quartz.Impl;
-using Quartz.Util;
 
 namespace Quartz.Tests.Unit.Impl.Calendar;
