+++ src/Quartz.Tests.Unit/Impl/Recurrence/RecurrenceRuleTest.cs
@@ -1,5 +1,4 @@
 using Quartz.Impl.Recurrence;
-using Quartz.Util;
 
 namespace Quartz.Tests.Unit.Impl.Recurrence;
