+++ src/Quartz.Tests.Unit/TestDates.cs
@@ -17,8 +17,6 @@
  */
 #endregion
 
-using Quartz.Util;
-
 namespace Quartz.Tests.Unit;
 
 /// <summary>
