+++ src/Quartz/Impl/AdoJobStore/CalendarIntervalTriggerPersistenceDelegate.cs
@@ -21,7 +21,6 @@
 
 using Quartz.Impl.Triggers;
 using Quartz.Extensibility;
-using Quartz.Util;
 
 namespace Quartz.Impl.AdoJobStore;
