+++ src/Quartz/Impl/AdoJobStore/DailyTimeIntervalTriggerPersistenceDelegate.cs
@@ -1,4 +1,4 @@
-﻿#region License
+#region License
 
 /*
  * All content copyright Marko Lahma, unless otherwise indicated. All rights reserved.
@@ -24,7 +24,6 @@ using System.Text;
 
 using Quartz.Impl.Triggers;
 using Quartz.Extensibility;
-using Quartz.Util;
 
 namespace Quartz.Impl.AdoJobStore;
