+++ src/Quartz/Impl/AdoJobStore/RecurrenceTriggerPersistenceDelegate.cs
@@ -1,7 +1,6 @@
 
 using Quartz.Impl.Triggers;
 using Quartz.Extensibility;
-using Quartz.Util;
 
 namespace Quartz.Impl.AdoJobStore;
