+++ src/Quartz/Impl/Calendar/MonthlyCalendar.cs
@@ -21,8 +21,6 @@
 
 using System.Runtime.Serialization;
 
-using Quartz.Util;
-
 namespace Quartz.Impl.Calendar;
 
 /// <summary>
